Managed to get the bounding box thing to work -- now enemies are killed when the laser hits them.
Need to find a way to delete the objects from the game without the game crashing

Also the bounding box of enemies is slightly too big to the left and need to adjust it such it wraps the sprite only.

Need to do more testing and clean up the code.